+++ src/main/java/backend/Race.java
@@ -1,7 +1,5 @@
 package backend;
 
-import frontend.GameEngine;
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
