+++ src/main/java/frontend/Manager.java
@@ -1,5 +1,6 @@
 package frontend;
 
+import backend.GameEngine;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
