+++ src/main/java/frontend/controllers/PrepareRaceController.java
@@ -4,7 +4,7 @@ import backend.Season;
 import backend.Setup;
 import backend.Strategy;
 import backend.Team;
-import frontend.GameEngine;
+import backend.GameEngine;
 
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
