+++ src/main/java/frontend/controllers/SelectEngineController.java
@@ -2,7 +2,7 @@ package frontend.controllers;
 
 import backend.Engine;
 import backend.Season;
-import frontend.GameEngine;
+import backend.GameEngine;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
