+++ src/main/java/frontend/controllers/SelectTeamMemberController.java
@@ -7,7 +7,7 @@ import backend.Season;
 import backend.Staff;
 import backend.Strategist;
 import backend.Team;
-import frontend.GameEngine;
+import backend.GameEngine;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
