+++ src/main/java/frontend/controllers/TransferOfferController.java
@@ -4,7 +4,7 @@ import backend.Driver;
 import backend.Season;
 import backend.Staff;
 import backend.Team;
-import frontend.GameEngine;
+import backend.GameEngine;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
